requirements.txt
streamlit==1.35.0
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.2.2
python-dotenv==1.0.1
bcrypt==4.1.3
Pillow==10.3.0
requests==2.31.0
urllib3==1.26.18
charset-normalizer==3.3.2
